Thy Warnings Conumed (MSVC Part2). Handle more level 4 warnings and put the libs on a separate project with warning level 0

+++ neo/d3xp/ai/AAS_routing.cpp
@@ -191,7 +191,7 @@ void idAASLocal::CalculateAreaTravelTimes()
 	}
 	
 	// RB: 64 bit fixes, changed unsigned int to ptrdiff_t
-	assert( ( ( ptrdiff_t ) bytePtr - ( ptrdiff_t ) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	assert( ( ( ptrdiff_t ) bytePtr - ( ptrdiff_t ) areaTravelTimes ) <= (unsigned)(numAreaTravelTimes * sizeof( unsigned short )) );
 	// RB end
 	
 }
